ai call FB/AI-call-intelligence-be/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pgvector==0.3.5

# Configuration
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP Client
httpx==0.27.2

# Background Tasks
celery[redis]==5.4.0

# AI / ML
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2
whisperx>=3.1.1
scikit-learn>=1.3.0
numpy>=1.24.0

# Report Generation
jinja2==3.1.4
weasyprint==62.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
aiofiles==24.1.0
